mvp-assessment/agent-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Developer assessment platform backend"
readme = "README.md"
authors = [
    { name = "Pieter Swanepoel", email = "130872729+PieterJDSw@users.noreply.github.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "json-repair>=0.52.4",
    "litellm>=1.79.3",
    "pika>=1.3.2",
    "prometheus-client>=0.25.0",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.11.10",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.24",
    "sqlalchemy>=2.0.49",
    "strands-agents>=1.0.0",
    "uvicorn>=0.44.0",
]

[build-system]
requires = ["uv_build>=0.11.4,<0.12.0"]
build-backend = "uv_build"
